Refactor batched vision cache modules
