+++ mlx_engine/model_kit/batched_vision/__init__.py
@@ -1,0 +1,3 @@
+from mlx_engine.model_kit.batched_vision.model_kit import BatchedVisionModelKit
+
+__all__ = ["BatchedVisionModelKit"]
