+++ mlx_engine/utils/eot_tokens.py
@@ -1,6 +1,6 @@
 from typing import Optional
 from mlx_engine.model_kit.batched_model_kit import BatchedModelKit
-from mlx_engine.model_kit.batched_vision_model_kit import BatchedVisionModelKit
+from mlx_engine.model_kit.batched_vision import BatchedVisionModelKit
 from mlx_engine.model_kit.model_kit import ModelKit
 from mlx_engine.vision_model_kit.vision_model_kit import VisionModelKit
